requirements.txt
djangorestframework
markdown
django-filter
social-auth-app-django